# Conflicts:
#	Jenkinsfile

+++ Jenkinsfile
@@ -8,24 +8,50 @@ pipeline{
             rollback = 'false'
         }
 
-
-          stages{
+        stages{
             stage('Clone Repo'){
                 steps{
-                    sh '''
-                    rm -rf Final-Project
-                    git clone https://github.com/makhdoomshabir/Final-Project.git
-                    cd Final-Project
-                    '''
-              }
-              }
-            
+                    script{
+                        if (env.rollback == 'false'){
 
+                            sh """
+                            rm -rf Final-Project
+                            git clone https://github.com/makhdoomshabir/Final-Project.git
+                            cd Final-Project
+                            export MYSQL_ROOT_PASSWORD=${MYSQL_ROOT_PASSWORD}
+                            export MYSQL_USER=${MYSQL_USER}
+                            export MYSQL_PASSWORD=${MYSQL_PASSWORD}
+                            export TEST_DATABASE_URI=${TEST_DATABASE_URI}
+                            export DATABASE_URI=${DATABASE_URI}
+                            export SECRET_KEY=${SECRET_KEY}
 
+                            sudo -E MYSQL_ROOT_PASSWORD=${MYSQL_ROOT_PASSWORD} MYSQL_USER=${MYSQL_USER} MYSQL_PASSWORD=${MYSQL_PASSWORD} DB_PASSWORD=${DB_PASSWORD} DATABASE_URI=${DATABASE_URI} SECRET_KEY=${SECRET_KEY} docker-compose build
+                            """
+
+                        }
+                    }
+                }
+            }
+            stage('Tag & Push Images'){
+                steps{
+                    script{
+                        if (env.rollback == 'false'){
+                            docker.withRegistry('https://registry.hub.docker.com', 'docker-credentials'){
+                                sh"""
+                                docker push krystalsimmonds/sfia-three-react:${env.app_version}
+                                docker push krystalsimmonds/sfia-three-spring:${env.app_version}
+                                docker push krystalsimmonds/mysql:5.7
+                                """
+                            }
+                        }
+                    }
+                }
+            }
             stage('Deploy App'){
                 steps{
-                    sh '''
+                    sh """
                     ssh ubuntu@10.0.2.114 <<EOF
+                    sudo rm -rf Final-Project
                     git clone https://github.com/makhdoomshabir/Final-Project.git
                     cd Final-Project
                     docker pull krystalsimmonds/sfia-three-react:${env.app_version}
@@ -34,22 +60,22 @@ pipeline{
 
                     docker-compose up -d
                     EOF
-                    '''
+                    """
                 }
             }
             stage('Configure kubectl'){
                 steps{
                     withAWS(credentials: 'aws-credentials', region: 'eu-west-2'){
-                    sh '''
+                    sh """
                     aws eks update-kubeconfig --name sfia-three
-                    '''
+                    """
                     }
                 }
             }
             stage('Deploy with k8s'){
                 steps{
                     withAWS(credentials: 'aws-credentials', region: 'eu-west-2'){
-                        sh '''
+                        sh """
                         rm -rf Final-Project
                         git clone -b DevOps https://github.com/makhdoomshabir/Final-Project.git
                         cd Final-Project/kubernetes
@@ -59,7 +85,7 @@ pipeline{
                         kubectl apply -f mysql-db.yaml
                         kubectl apply -f backend.yaml
                         kubectl apply -f frontend.yaml
-                        '''
+                        """
                     }
                 }
             }
